Clean up.

+++ Volund/src/Rendering/IndexBuffer/IndexBuffer.cpp
@@ -8,7 +8,7 @@
 
 namespace Volund
 {
-	Ref<IndexBuffer> IndexBuffer::Create(const uint32_t indices[], uint32_t count)
+	Ref<IndexBuffer> IndexBuffer::Create(const uint32_t indices[], uint64_t count)
 	{
 		switch (RenderingAPI::GetSelectedAPI())
 		{
